requirements.txt
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.11
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
uritemplate==4.2.0
urllib3==2.5.0
